+++ src/main/kotlin/com/zepben/evolve/services/network/translator/NetworkProtoToCim.kt
@@ -1142,6 +1142,10 @@ fun toCim(pb: PBFeeder, networkService: NetworkService): Feeder =
             networkService.resolveOrDeferReference(Resolvers.normalEnergizedLvFeeders(this), normalEnergizedLvFeederMRID)
         }
 
+        pb.currentlyEnergizedLvFeedersMRIDsList.forEach { currentEnergizedLvFeederMRID ->
+            networkService.resolveOrDeferReference(Resolvers.currentEnergizedLvFeeders(this), currentEnergizedLvFeederMRID)
+        }
+
         toCim(pb.ec, this, networkService)
     }
 
@@ -2778,6 +2782,10 @@ fun toCim(pb: PBLvFeeder, networkService: NetworkService): LvFeeder =
         networkService.resolveOrDeferReference(Resolvers.normalHeadTerminal(this), pb.normalHeadTerminalMRID)
         pb.normalEnergizingFeederMRIDsList.forEach { normalEnergizingFeederMRID ->
             networkService.resolveOrDeferReference(Resolvers.normalEnergizingFeeders(this), normalEnergizingFeederMRID)
+        }
+
+        pb.currentlyEnergizingFeederMRIDsList.forEach { currentEnergizingFeederMRID ->
+            networkService.resolveOrDeferReference(Resolvers.currentEnergizingFeeders(this), currentEnergizingFeederMRID)
         }
         toCim(pb.ec, this, networkService)
     }